requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pysubs2>=1.6.0
pyyaml>=6.0
openai>=1.0.0
yt-dlp>=2024.0.0
